backend/app/services/translator_aspose.py
# translator_aspose.py
import os
import requests
from pathlib import Path
from dotenv import load_dotenv

# Load .env robustly
THIS_FILE = Path(__file__).resolve()
PROJECT_ROOT = THIS_FILE.parents[2]
dotenv_path = PROJECT_ROOT / ".env"
if dotenv_path.exists():
    load_dotenv(dotenv_path)

# expected env vars: ASPOSE_CLIENT_ID, ASPOSE_CLIENT_SECRET, ASPOSE_BASE_URL (optional)
ASPOSE_CLIENT_ID = os.getenv("ASPOSE_CLIENT_ID", "")
ASPOSE_CLIENT_SECRET = os.getenv("ASPOSE_CLIENT_SECRET", "")
ASPOSE_BASE_URL = os.getenv("ASPOSE_BASE_URL", "https://api.aspose.cloud")  # 可能需要根据区域调整

def _raise_if_not_configured():
    if not ASPOSE_CLIENT_ID or not ASPOSE_CLIENT_SECRET:
        raise RuntimeError("Aspose 未配置。请在 .env 中设置 ASPOSE_CLIENT_ID 和 ASPOSE_CLIENT_SECRET")

def translate_doc(input_path: str, output_path: str, src_lang='zh', tgt_lang='en'):
    """
    Aspose Cloud 方案（示例/占位）
    - 本示例给出思路：上传文件 -> 使用 Aspose API 导出 HTML 或者直接替换文档中的文本
    - 推荐做法：先调用 Aspose 将 docx 转为 HTML（带全部 drawing 与样式），
      将 HTML 发送到翻译 API（DeepSeek/你选的），然后用 Aspose 把翻译的 HTML 转回 docx。
    - 下面仅是模板：请参考 Aspose Cloud 文档并替换为具体 API 路径/参数/鉴权方式（Aspose SDK 更易用）。
    """
    _raise_if_not_configured()
    # TODO: 使用 Aspose SDK 或 REST API 完成以下步骤：
    # 1) 获取 Access Token（如果你的使用方式需要）
    # 2) 上传 input_path 到 Aspose Cloud Storage 或直接调用转换接口（某些接口接受文件流）
    # 3) 调用 API 将 docx -> html（保留 drawing/media）
    # 4) 使用 DeepSeek 翻译返回的 HTML（同 HTML 方案）
    # 5) 将翻译后的 HTML 提交给 Aspose 转回 docx
    # 6) 下载最终 docx 到 output_path
    #
    # 下面是一个非常简化的伪代码结构（不可直接运行，供你替换为具体 API 调用）：
    raise NotImplementedError(
        "Aspose 方案尚未在此示例中实现。请参考 Aspose Cloud 文档并在本文件中实现上传/转换/下载逻辑。"
    )
