backend/requirements.txt
fastapi
uvicorn[standard]
celery
redis
python-docx
python-pptx
openpyxl
sqlalchemy
requests
python-multipart
beautifulsoup4
Pillow
python-dotenv
openai
lxml
langdetect
psycopg2-binary
chardet
python-jose[cryptography]
passlib
bcrypt
